+++ ambient_toolbox/tests/structure_validator/settings.py
@@ -4,11 +4,14 @@ from django.conf import settings
 TEST_STRUCTURE_VALIDATOR_FILE_ALLOWLIST = []
 _TEST_STRUCTURE_VALIDATOR_FILE_WHITELIST = []
 
+TEST_STRUCTURE_VALIDATOR_MISPLACED_TEST_FILE_ALLOWLIST = []
+_TEST_STRUCTURE_VALIDATOR_MISPLACED_TEST_FILE_WHITELIST = []
+
 
 @property
 def TEST_STRUCTURE_VALIDATOR_FILE_WHITELIST():  # noqa: N802
     """
-    The term "Whitelist" will be deprecated in 12.2 and move to "Allowlist".
+    The term "Whitelist" will be deprecated in 12.7 and move to "Allowlist".
     Until then, keep this for backwards compatibility but warn users about future deprecation.
     """
 
@@ -26,7 +29,7 @@ def TEST_STRUCTURE_VALIDATOR_FILE_WHITELIST():  # noqa: N802
 @TEST_STRUCTURE_VALIDATOR_FILE_WHITELIST.setter
 def TEST_STRUCTURE_VALIDATOR_FILE_WHITELIST(value):  # noqa: N802
     """
-    The term "Whitelist" will be deprecated in 12.2 and move to "Allowlist".
+    The term "Whitelist" will be deprecated in 12.7 and move to "Allowlist".
     Until then, keep this for backwards compatibility but warn users about future deprecation.
     """
 
@@ -43,6 +46,44 @@ def TEST_STRUCTURE_VALIDATOR_FILE_WHITELIST(value):  # noqa: N802
     settings._TEST_STRUCTURE_VALIDATOR_FILE_WHITELIST = value
 
 
+@property
+def TEST_STRUCTURE_VALIDATOR_MISPLACED_TEST_FILE_WHITELIST():  # noqa: N802
+    """
+    The term "Whitelist" will be deprecated in 12.7 and move to "Allowlist".
+    Until then, keep this for backwards compatibility but warn users about future deprecation.
+    """
+
+    import warnings
+
+    warnings.warn(
+        "The 'TEST_STRUCTURE_VALIDATOR_MISPLACED_TEST_FILE_WHITELIST' setting is deprecated and will be removed in a future version."
+        " Use 'TEST_STRUCTURE_VALIDATOR_MISPLACED_TEST_FILE_ALLOWLIST' instead.",
+        category=DeprecationWarning,
+        stacklevel=1,
+    )
+    return _TEST_STRUCTURE_VALIDATOR_MISPLACED_TEST_FILE_WHITELIST
+
+
+@TEST_STRUCTURE_VALIDATOR_MISPLACED_TEST_FILE_WHITELIST.setter
+def TEST_STRUCTURE_VALIDATOR_MISPLACED_TEST_FILE_WHITELIST(value):  # noqa: N802
+    """
+    The term "Whitelist" will be deprecated in 12.7 and move to "Allowlist".
+    Until then, keep this for backwards compatibility but warn users about future deprecation.
+    """
+
+    import warnings
+
+    from django.conf import settings
+
+    warnings.warn(
+        "The 'TEST_STRUCTURE_VALIDATOR_MISPLACED_TEST_FILE_WHITELIST' setting is deprecated and will be removed in a future version."
+        " Use 'TEST_STRUCTURE_VALIDATOR_MISPLACED_TEST_FILE_ALLOWLIST' instead.",
+        category=DeprecationWarning,
+        stacklevel=1,
+    )
+    settings._TEST_STRUCTURE_VALIDATOR_MISPLACED_TEST_FILE_WHITELIST = value
+
+
 try:
     TEST_STRUCTURE_VALIDATOR_BASE_DIR = settings.BASE_DIR
 except AttributeError:
@@ -51,4 +92,3 @@ except AttributeError:
 TEST_STRUCTURE_VALIDATOR_BASE_APP_NAME = "apps"
 TEST_STRUCTURE_VALIDATOR_APP_LIST = settings.INSTALLED_APPS
 TEST_STRUCTURE_VALIDATOR_IGNORED_DIRECTORY_LIST = []
-TEST_STRUCTURE_VALIDATOR_MISPLACED_TEST_FILE_WHITELIST = []